Refactor Sketcher.vue to integrate sketcherStore for state management, streamline door handle selection, and enhance parameter handling for openings. Added new input for 'mp' and updated form submission logic.
